Deleted calcRaceData.js and moved its contents to calculateValues.js
